Preferences UI separated from config manager
