+++ src/main/scala/org/fs/mael/ui/prefs/RichFieldEditorPreferencePage.scala
@@ -1,4 +1,4 @@
-package org.fs.mael.ui.components
+package org.fs.mael.ui.prefs
 
 import org.eclipse.jface.preference.FieldEditor
 import org.eclipse.jface.preference.FieldEditorPreferencePage
